LayoutComponent added to App, Header component added and Tab component added
LayoutComponent added so that in future react router dom can be added to main component and in App component redux and graphql can be added to manage state.

This will ensure that the codebase will be clean and easy to manage. Header component added so that if any routes added it will be easier to navigate.

bootstrap added as a main style framework

yarn add bootstrap popper.js jQuery

code refactored by removing div wrapper in all jsx file

webpack alias config added for api