unauthenticated vs authenticated root url
unauthenticated users get home#index whilst authenticated users get
app#show. Also up react router to v6 beta as that is the new way
forward. There is a placeholder Charts component and there seems to be
an issue with the active link for the default /app Add path in react.