[Round Trip Data] A start on importing user data via CSV
Created a test for data upload.
Some changes to the html for the Stats page, to allow for better targeting in tests.
Added an ugly upload button to Home.js
Created a support class for handling CSV uploads (and downloads, later).
Also created a sample csv file, which really ought to be replaced with some Factory.